pr review: [ON-2522] frontend changes
